cleanup: Start of no-implicit-bool compliance; use prefix-increment.
Semantics of postfix-increment are to return the value it had before the
increment, encouraging code like `f(..., b++)`, which is better written
as `f(..., b); ++b;` to avoid doing too many possibly mutating
operations in one statement/expression.

I've put one of the no-implicit-bool cleanups in this PR, because it's
the only one that requires `CMP_NULL` and an addition to the CI to
continue supporting C++ as a compile target for cmp.c (it does today,
and I don't want to break that). There are a number of implicit bool
conversions, some of which are probably bugs, so I'll send a separate PR
to fix those.